requirements.txt
beautifulsoup4
seaborn
matplotlib
pandas
requests
python-dotenv